\name{golubKEGGSplits}
\alias{golubKEGGSplits}
\docType{data}
\title{Examplar splitSet}
\description{
  This is a data object precomputed by \code{adSplit} for
  illustration.
}
\usage{data(golubKEGGSplits)}
\format{
  Annotation-driven split set
    holds 70 splits on 72 elements,
    scores range is: 3.382672 17.31385,
    empirical p-values range is: 0.005 0.955,
    q-value range is: 0.1633333 0.955.
}
\details{
  This object is generated by the following call:

  golubKEGGSplits <- adSplit(golubNorm, "KEGG", "hu6800", B=1000)

  where \code{golubNorm} is a normalized version of \code{Golub_Merge}
  from the \code{golubEsets} package.
}
\examples{
data(golubKEGGSplits)
}
\keyword{datasets}
