\name{print.splitSet}
\alias{print.splitSet}
\alias{print,splitSet-method}
\title{Print Summaries for splitSets}
\description{
  Prints information of splitSets.
}
\usage{
\method{print}{splitSet}(x, ...)
}
\arguments{
  \item{x}{the splitSet-object to be described.}
  \item{\dots}{not used.}
}
\author{Claudio Lottaz}
\seealso{\code{\link{adSplit}}}
\examples{
data(golubKEGGSplits)
print(golubKEGGSplits)
}
\keyword{internal}
