\name{print.split}
\alias{print.split}
\title{Print Information on Single Splits}
\description{
  Print information on a single split.
}
\usage{
print.split(x, ...)
}
\arguments{
  \item{x}{the split-object to be described.}
  \item{\dots}{not used.}
}
\author{Claudio Lottaz}
\seealso{\code{\link{diana2means}}}
\keyword{internal}
